variable type correction -> double to long double

+++ iterative_fib.c
@@ -40,7 +40,7 @@ int main(int argc, char *argv[]){
 
 	clock_t end_t = clock();
 
-	double time_taken = ((long double)(end_t - start_t)) / CLOCKS_PER_SEC;
+	long double time_taken = ((long double)(end_t - start_t)) / CLOCKS_PER_SEC;
 
 	FILE *file;
 	file = fopen("output/results.csv", "a");
@@ -50,7 +50,7 @@ int main(int argc, char *argv[]){
 		return 0;
 	}
 
-	fprintf(file, "iterative,%d,%lf\n", n, time_taken);
+	fprintf(file, "iterative,%d,%Lf\n", n, time_taken);
 	
 	fclose(file);
